refactor(learning-scenarios): address rlaerm review
- Move sendLearningScenarioPreviewMessageAction into the existing
  (authed)/(chat-bot)/learning-scenarios/actions.ts; server actions belong
  with the route, not under /api.
- Extract streamLearningScenarioReply (shared-chat/learning-scenario-stream.ts)
  so the streaming pipeline (RAG retrieval, system-prompt build, message
  pruning, image formatting, billed streaming, RabbitMQ event) lives in one
  place. Both shared-chat-service (student-facing) and the new
  learning-scenario-preview-service (teacher preview) now go through it.
- Move the preview service alongside its sibling under api/shared-chat/.
- Drop the duplicate page-frame in learning-scenario-preview-chat.tsx:
  the route already wraps it in DefaultPageLayout, so h-full is correct
  and the inner max-w-5xl belongs to the layout, not the component.
- Remove the now-empty api/learning-scenario-preview/ folder.
