+++ apps/chat-bot/src/app/(authed)/(chat-bot)/learning-scenarios/actions.ts
@@ -1,6 +1,8 @@
 'use server';
 
 import { requireAuth } from '@/auth/requireAuth';
+import { userHasCompletedTraining } from '@/auth/utils';
+import { checkProductAccess } from '@/utils/vidis/access';
 import { runServerAction } from '@shared/actions/run-server-action';
 import {
   createNewLearningScenario,
@@ -9,6 +11,10 @@ import {
   downloadFileFromLearningScenario,
   linkFileToLearningScenario,
 } from '@shared/learning-scenarios/learning-scenario-service';
+import { sendLearningScenarioPreviewMessage } from '@/app/api/shared-chat/learning-scenario-preview-service';
+import { ChatMessage, SendMessageResult } from '@/types/chat';
+
+export type { ChatMessage, SendMessageResult } from '@/types/chat';
 
 export async function deleteLearningScenarioAction({ id }: { id: string }) {
   const { user } = await requireAuth();
@@ -68,5 +74,39 @@ export async function downloadFileFromLearningScenarioAction({
     learningScenarioId,
     fileId,
     user,
+  });
+}
+
+export async function sendLearningScenarioPreviewMessageAction({
+  previewSessionId,
+  learningScenarioId,
+  messages,
+  modelId,
+}: {
+  previewSessionId: string;
+  learningScenarioId: string;
+  messages: ChatMessage[];
+  modelId: string;
+}): Promise<SendMessageResult> {
+  const [{ user, federalState }, hasCompletedTraining] = await Promise.all([
+    requireAuth(),
+    userHasCompletedTraining(),
+  ]);
+  const userAndContext = {
+    ...user,
+    federalState,
+  };
+  const productAccess = checkProductAccess({ ...userAndContext, hasCompletedTraining });
+
+  if (!productAccess.hasAccess) {
+    throw new Error(productAccess.errorType);
+  }
+
+  return sendLearningScenarioPreviewMessage({
+    previewSessionId,
+    learningScenarioId,
+    messages,
+    modelId,
+    user: userAndContext,
   });
 }
