+++ apps/chat-bot/src/app/api/shared-chat/shared-chat-service.ts
@@ -1,10 +1,4 @@
-import {
-  generateTextStreamWithBilling,
-  type Message as AiCoreMessage,
-  TokenPointsExceededError,
-  SharedChatExpiredError,
-} from '@ais-chat/ai-core';
-import { createTextStream } from '@/utils/streaming';
+import { TokenPointsExceededError, SharedChatExpiredError } from '@ais-chat/ai-core';
 import { getUserAndContextByUserId } from '@/auth/utils';
 import { checkProductAccess } from '@/utils/vidis/access';
 import {
@@ -17,39 +11,10 @@ import {
   dbGetLearningScenarioByIdAndInviteCode,
   dbUpdateTokenUsageBySharedLearningScenarioId,
 } from '@shared/db/functions/learning-scenario';
-import { dbGetRelatedSharedChatFiles } from '@shared/db/functions/files';
 import { sendRabbitmqEvent } from '@/rabbitmq/send';
-import { constructNewMessageEvent } from '@/rabbitmq/events/new-message';
 import { constructTokenBudgetExceededEvent } from '@/rabbitmq/events/budget-exceeded';
-import { constructLearningScenarioSystemPrompt } from './system-prompt';
-import { formatMessagesWithImages, limitChatHistory } from '../chat/utils';
-import { retrieveChunks } from '../rag/rag-service';
-import { logError } from '@shared/logging';
-import {
-  KEEP_FIRST_MESSAGES,
-  KEEP_RECENT_MESSAGES,
-  TOTAL_CHAT_LENGTH_LIMIT,
-} from '@/configuration-text-inputs/const';
 import { ChatMessage, SendMessageResult, createErrorResult } from '@/types/chat';
-import { extractImagesAndUrl } from '../file-operations/preprocess-image';
-import { ingestWebContent } from '../rag/ingestWebContent';
-
-/**
- * Converts frontend messages to ai-core message format
- */
-function convertToAiCoreMessages(systemPrompt: string, messages: ChatMessage[]): AiCoreMessage[] {
-  const result: AiCoreMessage[] = [{ role: 'system', content: systemPrompt }];
-
-  for (const msg of messages) {
-    if (msg.role === 'system') continue;
-    result.push({
-      role: msg.role === 'user' ? 'user' : 'assistant',
-      content: msg.content,
-    });
-  }
-
-  return result;
-}
+import { streamLearningScenarioReply } from './learning-scenario-stream';
 
 /**
  * Server Action to send a shared chat (learning scenario) message and stream the response.
@@ -126,102 +91,23 @@ export async function sendSharedChatMessage({
     return createErrorResult(new TokenPointsExceededError());
   }
 
-  // Get related files and web sources
-  const relatedFileEntities = await dbGetRelatedSharedChatFiles(sharedChat.id);
-  const urls = sharedChat.attachedLinks.filter((l) => l !== '');
-  const { processedUrls } = await ingestWebContent({
-    urls,
-    federalStateId: teacherUserAndContext.federalState.id,
-  });
-
-  const chunks = await retrieveChunks({
+  return streamLearningScenarioReply({
+    learningScenario: sharedChat,
+    user: teacherUserAndContext,
     messages,
-    federalStateId: teacherUserAndContext.federalState.id,
-    relatedFileEntities,
-    sourceUrls: processedUrls,
+    model: definedModel,
+    apiKeyId,
+    eventAnonymous: true,
+    logTag: 'shared chat',
+    onUsage: async ({ promptTokens, completionTokens, costsInCent }) => {
+      await dbUpdateTokenUsageBySharedLearningScenarioId({
+        modelId: definedModel.id,
+        completionTokens,
+        promptTokens,
+        learningScenarioId: sharedChat.id,
+        userId: teacherUserAndContext.id,
+        costsInCent,
+      });
+    },
   });
-
-  // Build system prompt
-  const systemPrompt = constructLearningScenarioSystemPrompt({
-    sharedChat,
-    chunks,
-  });
-
-  // Prune messages
-  const prunedMessages = limitChatHistory({
-    messages: messages.map((m) => ({ id: m.id, role: m.role, content: m.content })),
-    limitRecent: KEEP_RECENT_MESSAGES,
-    limitFirst: KEEP_FIRST_MESSAGES,
-    characterLimit: TOTAL_CHAT_LENGTH_LIMIT,
-  });
-
-  // Check if the model supports images based on supportedImageFormats
-  const modelSupportsImages =
-    definedModel.supportedImageFormats !== null && definedModel.supportedImageFormats.length > 0;
-
-  // attach the image url to each of the image files within relatedFileEntities
-  const extractedImages = await extractImagesAndUrl(relatedFileEntities);
-
-  // Format messages with images if the model supports vision
-  const messagesWithImages = formatMessagesWithImages(
-    prunedMessages,
-    extractedImages,
-    modelSupportsImages,
-  );
-
-  // Convert to ai-core format
-  const aiCoreMessages = convertToAiCoreMessages(systemPrompt, messagesWithImages);
-
-  // Create native stream
-  const { stream, update, done, error: streamError } = createTextStream();
-  const assistantMessageId = crypto.randomUUID();
-
-  // Start streaming in the background
-  void (async () => {
-    try {
-      const textStream = generateTextStreamWithBilling(
-        definedModel.id,
-        aiCoreMessages,
-        apiKeyId,
-        async ({ usage, priceInCents }) => {
-          const { promptTokens, completionTokens } = usage;
-
-          await dbUpdateTokenUsageBySharedLearningScenarioId({
-            modelId: definedModel.id,
-            completionTokens,
-            promptTokens,
-            learningScenarioId: sharedChat.id,
-            userId: teacherUserAndContext.id,
-            costsInCent: priceInCents,
-          });
-
-          await sendRabbitmqEvent(
-            constructNewMessageEvent({
-              user: teacherUserAndContext,
-              provider: definedModel.provider,
-              promptTokens,
-              completionTokens,
-              costsInCent: priceInCents,
-              anonymous: true,
-              sharedChat,
-            }),
-          );
-        },
-      );
-
-      for await (const chunk of textStream) {
-        update(chunk);
-      }
-
-      done();
-    } catch (error) {
-      logError('Error during shared chat streaming:', error);
-      streamError(error instanceof Error ? error : new Error('Unknown error'));
-    }
-  })();
-
-  return {
-    stream,
-    messageId: assistantMessageId,
-  };
 }
