+++ apps/chat-bot/src/components/chat/learning-scenario-preview-chat.tsx
@@ -59,9 +59,9 @@ export default function LearningScenarioPreviewChat({
   const isLoading = status === 'submitted';
 
   return (
-    <div className="flex h-dvh min-h-0 w-full flex-col overflow-hidden">
+    <div className="flex h-full w-full flex-col overflow-hidden">
       <div ref={containerRef} className="relative flex min-h-0 flex-1 flex-col items-center w-full">
-        <div ref={scrollRef} className="min-h-0 w-full flex-1 max-w-5xl overflow-y-auto p-4 pb-20">
+        <div ref={scrollRef} className="min-h-0 w-full flex-1 overflow-y-auto pb-20">
           {learningScenario.studentExercise !== undefined &&
             learningScenario.studentExercise.trim() !== '' && (
               <FloatingText
@@ -93,7 +93,7 @@ export default function LearningScenarioPreviewChat({
           )}
           {error && <ErrorChatPlaceholder error={error} handleReload={handleReload} />}
         </div>
-        <div className="w-full max-w-5xl shrink-0 mx-auto px-4 pb-4">
+        <div className="w-full shrink-0">
           {dialogStarted && (
             <div className="flex flex-col">
               <ChatInputBox
